Generalize TestInitialize/Cleanup methods.
