+++ src/Tests/SenseNet.Tests.Core/Extensions.cs
@@ -1,10 +1,52 @@
-﻿using SenseNet.ContentRepository;
+﻿using System.Linq;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using SenseNet.ContentRepository;
 using SenseNet.ContentRepository.Storage;
+using SenseNet.Diagnostics;
+using SenseNet.Testing;
 
 namespace SenseNet.Tests.Core
 {
     public static class Extensions
     {
+        public static void StartTest(this TestContext testContext, bool traceToFile = false)
+        {
+            if (traceToFile)
+            {
+                var tracers = SnTrace.SnTracers.ToArray();
+                testContext.Properties["SnTrace.Operation.Writers"] = tracers;
+                if (!tracers.Any(x => x is SnFileSystemTracer))
+                    SnTrace.SnTracers.Add(new SnFileSystemTracer());
+            }
+            StartTestPrivate(testContext);
+        }
+        private static void StartTestPrivate(TestContext testContext)
+        {
+            using (new Swindler<bool>(true, () => SnTrace.Test.Enabled, x => SnTrace.Test.Enabled = x))
+                testContext.Properties["SnTrace.Operation"] =
+                    SnTrace.Test.StartOperation("TESTMETHOD: " + testContext.TestName);
+        }
+        public static void FinishTestTest(this TestContext testContext)
+        {
+            using (new Swindler<bool>(true, () => SnTrace.Test.Enabled, x => SnTrace.Test.Enabled = x))
+            {
+                var op = (SnTrace.Operation)testContext.Properties["SnTrace.Operation"];
+                SnTrace.Test.Write("TESTMETHOD: {0}: {1}", testContext.TestName, testContext.CurrentTestOutcome);
+                if (op != null)
+                {
+                    op.Successful = true;
+                    op.Dispose();
+                }
+                SnTrace.Flush();
+            }
+            var originalTracers = (ISnTracer[])testContext.Properties["SnTrace.Operation.Writers"];
+            if (originalTracers != null)
+            {
+                SnTrace.SnTracers.Clear();
+                SnTrace.SnTracers.AddRange(originalTracers);
+            }
+        }
+
         /// <summary>
         /// Returns with a new child of the given parent by the specified content type and name.
         /// The output and return values are reference equal.
